I changed line 835 from board.position(fen) --> chessboard.position(fen)
line 9 and line 11 are duplicated I put # in line 9
On line 556 there is a missing part of the line.
I continued the line as far as I understood (not sure if it was correct though).
    elif self.board_index[self.get_ref_from_index(two_ahead_index)]["type"] != None:

from line 685 - it appears that the whole code is duplicated here, so I turned the rest of the code to a comment(""")
until line 1369.